+++ PagaLaEscuela/Controllers/IntegracionesClubPago/GenerarReferenciaIndiController.cs
@@ -194,7 +194,7 @@ namespace PagaLaEscuela.Controllers.IntegracionesClubPago
                 {
                     try
                     {
-                        int Reference = int.Parse(generarFormaPago.Reference);
+                        Int64 Reference = Int64.Parse(generarFormaPago.Reference);
 
                         if (generarFormaPago.Reference.Length == 13)
                         {
