+++ PagaLaEscuela/Controllers/IntegracionesPraga/GenerarLigaIndiController.cs
@@ -249,7 +249,7 @@ namespace PagaLaEscuela.Controllers.IntegracionesPraga
                 {
                     try
                     {
-                        int Reference = int.Parse(generarFormaPago.Reference);
+                        Int64 Reference = Int64.Parse(generarFormaPago.Reference);
 
                         if (generarFormaPago.Reference.Length == 13)
                         {
